requirements.txt
streamlit
pytube
openai
openai-whisper
python-dotenv